implemented downlaod API definition, collapse/expand menu items
