+++ src/components/apis/api-details-page/ko/runtime/api-details-page.html
@@ -10,8 +10,8 @@
     </div>
     <div class="page-menu">
         <div class="list-container">
-            <div class="items">
-                <api-list-dropdown></api-list-dropdown>
+            <div class="items" data-bind="{css: {'full': api().type !== 'websocket' || api().type !== 'graphql'}}">
+                <api-list-dropdown params="{ allowSelection: true }"></api-list-dropdown>
                 <input type="search" class="form-control form-control-light" aria-label="Search this page"
                     placeholder="Search this page" spellcheck="false" data-bind="textInput: pattern" />
                 <!-- ko if: versionApis().length > 0 -->
@@ -26,51 +26,83 @@
                 </label>
                 <!-- /ko -->
                 <hr>
-                <div class="menu-list-heading">
-                    About
-                </div>
-                <div class="menu-list">
-                    <!-- ko foreach: { data: staticSelectableMenuItems, as: 'menuItem' } -->
-                    <div class="menu-list-item" role="option"
-                        data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
-                        <span data-bind="text: menuItem.displayName"></span>
+                <details open class="details-styled">
+                    <summary aria-label="About">
+                        <div class="menu-list-heading">
+                            About
+                            <i class="icon-emb icon-emb-chevron-down"></i>
+                        </div>
+                    </summary>
+                    <div class="menu-list">
+                        <!-- ko foreach: { data: staticSelectableMenuItems, as: 'menuItem' } -->
+                        <div class="menu-list-item" role="option"
+                            data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
+                            <span data-bind="text: menuItem.displayName"></span>
+                        </div>
+                        <!-- /ko -->
                     </div>
-                    <!-- /ko -->
-                </div>
+
+                </details>
                 <!-- ko if: wikiDocumentationMenuItems()?.length > 0 -->
                 <hr>
-                <div class="menu-list-heading">
-                    Documentatons
-                </div>
-                <div class="menu-list">
-                    <!-- ko foreach: { data: wikiDocumentationMenuItems(), as: 'menuItem' } -->
-                    <div class="menu-list-item" role="option"
-                        data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
-                        <span data-bind="text: menuItem.displayName"></span>
+                <details open class="details-styled">
+                    <summary aria-label="Documentations">
+                        <div class="menu-list-heading">
+                            Documentatons
+                            <i class="icon-emb icon-emb-chevron-down"></i>
+                        </div>
+                    </summary>
+                    <div class="menu-list">
+                        <!-- ko foreach: { data: wikiDocumentationMenuItems(), as: 'menuItem' } -->
+                        <div class="menu-list-item" role="option"
+                            data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
+                            <span data-bind="text: menuItem.displayName"></span>
+                        </div>
+                        <!-- /ko -->
                     </div>
-                    <!-- /ko -->
-                </div>
+                </details>
                 <!-- /ko -->
 
                 <!-- ko if: operationsMenuItems()?.length > 0 -->
                 <hr>
-                <div class="menu-list-heading">
-                    Operations
-                </div>
-                <div class="menu-list">
-                    <!-- ko foreach: { data: operationsMenuItems(), as: 'menuItem' } -->
-                    <div class="menu-list-item" role="option"
-                        data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
-                        <span data-bind="text: menuItem.displayName"></span>
+                <details open class="details-styled">
+                    <summary aria-label="Operations">
+                        <div class="menu-list-heading">
+                            Operations
+                            <i class="icon-emb icon-emb-chevron-down"></i>
+                        </div>
+                    </summary>
+                    <div class="menu-list">
+                        <!-- ko foreach: { data: operationsMenuItems(), as: 'menuItem' } -->
+                        <div class="menu-list-item" role="option"
+                            data-bind="activate: $component.selectMenuItem, css: { 'selected': $component.selectedMenuItem() == menuItem }">
+                            <span data-bind="text: menuItem.displayName"></span>
+                        </div>
+                        <!-- /ko -->
                     </div>
-                    <!-- /ko -->
-                </div>
+                </details>
                 <!-- /ko -->
             </div>
-            <div class="sticky-bottom">
+            <!-- ko if: api().type !== 'websocket' && api().type !== 'graphql' -->
+            <div class="bottom-item">
                 <hr>
-                <span>Download API definition</span>
+                <label class="select-badge">
+                    <select id="selectedDefinition"  data-bind="value: selectedDefinition">
+                        <option value="" disabled="disabled" selected="selected">Download API definition</option>
+                        <option value="openapi">Open API 3 (YAML)</option>
+                        <option value="openapi+json">Open API 3 (JSON)</option>
+                        <option value="swagger"> Open API 2 (JSON)</option>
+                        <option value="wadl">WADL</option>
+                        <!-- ko if: api().type === 'soap' -->
+                        <option value="wsdl">WSDL</option>
+                        <!-- /ko -->
+                    </select>
+                    <span class="arrow flex align-items-center justify-content-center">
+                        <i class="icon-emb icon-emb-chevron-down"></i>
+                    </span>
+                </label>
             </div>
+            <!-- /ko -->
 
         </div>
     </div>
